+++ src/components/cms/CMSProductGridBase.tsx
@@ -1,5 +1,15 @@
 import React from 'react'
-import type { CMSProductGridItem } from './CMSProductGridOld'
+import type { CMSProductGridItem } from './CMSProductGridItem'
+
+// Map numeric values to valid Tailwind class names
+const COLUMN_MAP: Record<number, string> = {
+	1: 'grid-cols-1',
+	2: 'grid-cols-2',
+	3: 'grid-cols-3',
+	4: 'grid-cols-4',
+	5: 'grid-cols-5',
+	6: 'grid-cols-6',
+}
 
 export interface ProductGridBaseProps {
 	items: CMSProductGridItem[]
@@ -34,60 +44,62 @@ export const ProductGridBase: React.FC<ProductGridBaseProps> = ({
 		return <div className="py-12 text-center text-gray-500">No products found matching your criteria.</div>
 	}
 
-	const gridCols = `
-    grid-cols-${columnsMobile} 
-    md:grid-cols-${columnsTablet} 
-    lg:grid-cols-${columnsDesktop}
-  `.replace(/grid-cols-\d+/g, (match) => match.replace('grid-cols-', ''))
+	// Safely resolve class names using the map
+	// Fallback to 'grid-cols-1' if the number is outside our map (e.g., 7+)
+	const mobileClass = COLUMN_MAP[columnsMobile] || 'grid-cols-1'
+	const tabletClass = COLUMN_MAP[columnsTablet] || 'grid-cols-2'
+	const desktopClass = COLUMN_MAP[columnsDesktop] || 'grid-cols-3'
 
 	return (
 		<div className="py-12 px-6 max-w-7xl mx-auto">
-			<div className={`grid gap-8 ${gridCols}`}>
+			{/* 
+			 Tailwind classes are now static strings derived from the map.
+			 This ensures the compiler generates the correct CSS.
+			*/}
+			<div className={`grid gap-8 ${mobileClass} md:${tabletClass} lg:${desktopClass}`}>
 				{items.map((product) => (
 					<div
 						key={product.id}
-						className="group relative bg-white rounded-lg shadow-sm hover:shadow-lg transition-shadow duration-300 overflow-hidden border border-gray-100"
+						className="group relative bg-white rounded-lg shadow-sm hover:shadow-lg transition-shadow duration-300 overflow-hidden border border-gray-100 flex flex-col"
 					>
-						{/* Badge */}
-
-						{/*product.badge && (
-                            <div
-                                className="absolute top-3 left-3 z-10 px-3 py-1 text-xs font-bold text-white rounded-full"
-                                style={{ backgroundColor: product.badgeColor || '#ef4444' }}
-                            >
-                                {product.badge}
-                            </div>
-                        )*/}
-
-						{/* Image */}
+						{/* Image Container */}
 						<div className="relative aspect-square overflow-hidden bg-gray-100">
-							<img
-								src={product.image}
-								alt={product.title}
-								className="w-full h-full object-cover transform group-hover:scale-105 transition-transform duration-500"
-							/>
+							{product.image ? (
+								<img
+									src={product.image}
+									alt={product.title}
+									className="w-full h-full object-cover transform group-hover:scale-105 transition-transform duration-500"
+								/>
+							) : (
+								<div className="w-full h-full flex items-center justify-center text-gray-400">No Image</div>
+							)}
 
 							{/* Quick Add Button */}
 							{showQuickAdd && (
-								<button className="absolute bottom-4 left-1/2 -translate-x-1/2 translate-y-12 group-hover:translate-y-0 opacity-0 group-hover:opacity-100 transition-all duration-300 bg-white text-gray-900 px-4 py-2 rounded-md font-medium shadow-md hover:bg-gray-50">
+								<button className="absolute bottom-4 left-1/2 -translate-x-1/2 translate-y-12 group-hover:translate-y-0 opacity-0 group-hover:opacity-100 transition-all duration-300 bg-white text-gray-900 px-4 py-2 rounded-md font-medium shadow-md hover:bg-gray-50 whitespace-nowrap">
 									Quick Add
 								</button>
 							)}
 						</div>
 
 						{/* Content */}
-						<div className="p-4">
-							{showVendor && product.pubkey && <p className="text-xs text-gray-500 uppercase tracking-wide mb-1">{product.pubkey}</p>}
-							<h3 className="text-lg font-semibold text-gray-900 mb-2 truncate">
+						<div className="p-4 flex flex-col flex-1">
+							{showVendor && product.pubkey && (
+								<p className="text-xs text-gray-500 uppercase tracking-wide mb-1 truncate">
+									{product.pubkey.slice(0, 6)}...{product.pubkey.slice(-4)}
+								</p>
+							)}
+							<h3 className="text-lg font-semibold text-gray-900 mb-2 line-clamp-2">
 								<a href={`/product/${product.id}`} className="hover:text-orange-600 transition-colors">
 									{product.title}
 								</a>
 							</h3>
 
-							<p className="text-lg font-bold text-gray-900">
-								{product.currency || '$'}
-								{product.price}
-							</p>
+							<div className="mt-auto">
+								<p className="text-lg font-bold text-gray-900">
+									{product.currency || 'SATS'} {product.price}
+								</p>
+							</div>
 						</div>
 					</div>
 				))}
